+++ frontend/static/src/user-profile-components/ProfileDetail.js
