+++ frontend/static/src/user-profile-components/UserRecipeList.js
@@ -47,7 +47,6 @@ class UserRecipeList extends Component {
           'Authorization': `Token ${JSON.parse(localStorage.getItem('current-user')).key}`
         }
       })
-
     .then(response => this.setState({recipes: response.data}))
     .catch(err => console.log(err));
