+++ frontend/static/src/utils/makeAPICall.js
@@ -2,7 +2,7 @@ import axios from "axios";
 
 const BASE_URL = process.env.REACT_APP_BASE_URL;
 
-export const makeAPICall = (endpoint) => {
+export const GetAPICall = (endpoint) => {
     return axios.get(`${BASE_URL}/api/v1/${endpoint}/`, {
         headers: {
             'Authorization': `Token ${JSON.parse(localStorage.getItem('current-user')).key}`
@@ -11,4 +11,12 @@ export const makeAPICall = (endpoint) => {
     // Because we return axios.get() we specify a .then() and .catch() when makeAPICall() is called.
 
 }
-export default makeAPICall;
+
+export const PostAPICall = (endpoint, data) => {
+    return axios.post(`${BASE_URL}/api/v1/${endpoint}`, data, {
+        headers: {
+            'Authorization': `Token ${JSON.parse(localStorage.getItem('current-user')).key}`
+        }
+    });
+}
+export default GetAPICall
